refactor return steps to separate component
